pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vortex-server-python"
version = "0.1.0"
description = "Perspective view server — Solace, NATS, WebSocket ingest"
requires-python = ">=3.11"
dependencies = [
    "perspective-python>=3.4",
    "tornado>=6.4",
    "nats-py[nkeys]>=2.9",
    "solace-pubsubplus>=1.8",
    "websockets>=12.0",
    "orjson>=3.10",
    "pydantic-settings>=2.4",
    "pyyaml>=6.0",
    "pyarrow>=16.0",
    "pymongo>=4.8",
    "flask>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
vortex-server = "vortex.server:run"
vortex-admin = "vortex.admin.app:run"

[tool.hatch.build.targets.wheel]
packages = ["vortex"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
